M-multi-species/scripts/20-new-sparse-tensor-decompositon-model/pyproject.toml
[project]
name = "new-tensor"
version = "0.1.0"
description = "Sparse tensor decomposition for multi-species time series molecular data"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "tensorly>=0.8.0",
    "optuna>=2.10.0",
    "typer[all]>=0.7.0",
    "matplotlib>=3.5.0",
    "pyyaml>=5.4.0",
    "click>=7.0.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
new-tensor = "new_tensor.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["new_tensor*"]

[tool.setuptools.package-data]
"new_tensor" = ["*.yaml"]

[tool.black]
line-length = 88
target-version = ['py38']
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
